moved extra functions from wavelet_new.cpp to separate file
