+++ examples/transform_examples/dwt_reducer.cpp
@@ -10,6 +10,7 @@
 
 #include "assets/helpers.cpp"
 
+//#include "transform/helper_functions.cpp" // TODO remove
 
 #include <vector>
 #include <iostream>
