+++ examples/transform_examples/wavelet_example.cpp
@@ -9,6 +9,8 @@
 #include "assets/helpers.cpp"
 #include "modules/utils/visualizer.hpp"
 
+//#include "transform/helper_functions.cpp" // TODO remove
+
 int main() {
 
     //std::vector<double> data = {0, 0, 0, 0, 1, 0, 0, 0, 0, 0};
